+++ mol_gnn/utils/__init__.py
@@ -1,3 +1,1 @@
-from .mixins import ReprMixin
-from .registry import ClassRegistry
 from .utils import *
